refactor: modularize control panel tabs
